requirements.txt
pandas
pandas_datareader
numpy
tensorflow
matplotlib.pyplot
keras
